requirements.txt
Flask==1.1.2
requests==2.25.1
mysql-connector-python==8.0.23
python-decouple==3.4
python-dotenv==0.15.0
azure-functions==1.7.0
